requirements.txt
numpy
opencv-python
Pillow
vk_api
requests